Fixes for the story parser to put in a timeout and chapter limit. Added more meta-harvesting from ff.net stories. Multi-chapter ff.net stories are not yet working, except for the first chapter, but progress is being made.
Also minor tweaks to the help icon style and link_to_help code, to get rid of the underline and tweak the appearance of the icon, and to improve the display names of help files. 

Also, migration for email in feedback model to complete Zooey's earlier commit and shush up the tests, and committed schema.rb. Please rake db:migrate or rake db:seed for sqlite users!