Made anonymity/unrevealed state of items in a collection changeable on an item-by-item basis, instead of only by the collection as a whole.
The recipient on a story should be notified whenever a story is revealed, but not repeatedly (ie, if a story was revealed, and then the whole collection was made unrevealed, the recipient should only receive the notification the first time).

Requires rake db:migrate and rake After:update_collection_items to update existing unrevealed/anonymous collection items.