Migrations for the new collections code. Adds the following classes:
Prompt -- can be either an offer or a request. Holds user-submitted information including a url, description, and a tagset. 

PromptRestriction -- holds a set of rules which can be used to limit either offers or requests. This is set up by a challenge moderator. Used heavily in prompt validation. 

ChallengeSignup -- a container model to group a number of prompts for a single user. 

GiftExchange -- a specific type of challenge which involves users posting requests and offers and getting matched up with each other. (Future challenges can be coded to have entirely different behavior.)