Improvements to layout of tag checkboxes for challenges in the prompt form.
New coding tool: if you create a div anywhere in the front end and give it a css class of "hideme" it will automatically be hidden on load. (You then still have to write code to show it, of course. :) See the prompt form for an example of usage. 

Fix for issue 1369 (dates not getting loaded correctly into challenge page, was not being used as form helper, grr). 

Also fixes from lim for weird-looking buttons in collection participant index and some other issues.