added mocking for testing (requires mocha gem)
mocked out story parser tests that seemed to be mostly parsing, and
didn't actually need to test connectivity to the source. 

Created mock stories, which have all the surrounding html cruft, but 
only the first and last few lines of the actual stories.

reduced run time for story_parser_test.rb from 146 seconds to 65 seconds
on my computer. Still not great, but better.