Bugfixes and a bit of refactoring of CSS sanitizing.
Wizard elements should now also be sanitized correctly, and error messages should be helpful. If a skin being created/edited in the wizard has errors, it should return to the wizard form and not the custom-css form. 

Added autocomplete to the wizard form, and a possibly helpful autocomplete widget on the custom-css form -- opinions welcome, is easy to yank. 

Tidied up minor things on skins forms to make them more consistent with each other/other forms.