RAILS 3 upgrade in progress!
Removing old acts_as_authenable plugin and replacing with Authlogic gem. Not everything has been updated yet, but it should be possible to log in both as a user and as an admin, and "remember_me" should work for users. 

OpenID is NOT working currently and has been stripped out for now.

Run bundle install and db:migrate after this update!