Added the "translate_string" and shortcut "ts" option -- should now be usable in all views, controllers, models, and mailers. Have changed a handful of uses of "t" to "ts" to test and demonstrate.
Also fixed a bunch more html_safe issues. 

Installed the prototype_legacy_helper plugin as a temporary measure to get observe_field back. Ideally we will want to replace this with an unobtrusive javascript way of doing the same thing, but while the spirit was willing, the brain was weak. :P

Password reset back end should be functional now although this can't be tested until we have mailers working. Account activation not done yet. 

The post new form should now come up and allow posting of stories, although it is chock-full of html_safe problems that we need to investigate further.