OMG.
Huge overhaul of sanitizing and parsing. We now save the sanitized and parsed versions in the database instead of re-sanitizing and parsing every time. Sanitized versions are tracked, so if we upgrade the sanitizer significantly, fields will automatically sanitize and save themselves as they go. 

Sanitizing and parsing is now being done via Nokogiri and the Sanitize gem, rather than the built-in Rails tokenizing sanitizer. 

The sanitizing functions have been simplified. Stripping images and various attributes at display-time can now be done with fast regexps because the content out of the database is guaranteed to be well-formed xhtml thanks to Nokogiri.

Titles do not allow HTML and do not need to be sanitized or escaped -- have just yanked all the sanitize calls on them. For safety, just do not mark them html_safe. 

To be done: more testing and tweaking of the paragraph/linebreak conversion, and figuring out whether to convert them back for editing.