switched to a different tolk fork so rake tasks don't get deprecation warnings.
the schema.rb was out of sync with the migrations, so I have reverted it to exactly as it exists on beta, and all migrations which have been made on beta have been removed.

the old tolk css link had been removed earlier as the plugin no longer exists. have replaced it with the current versions.