Yikes, fix for bug where apparently this kind of scope doesn't work:
# Use the current user to determine what works are visible
  scope :visible, visible_to_user(User.current_user)

User.current_user seems to be passed in as nil (which is better than the alternative because it causes us to err on the side of privacy). 

If we have visibility bugs this is probably going to be the source, and some more manual/cucumber testing of locked things would be good. :O