Fixes issue 2052: posting to a subcollection (eg like Yuletide2010) would fail because the collection model validations were being run even though the collection itself was unchanged.
Made the validations run specifically on save, and now the shiny new test passes. :D 

Also added a few more convenience steps for testing:

I should see "x" in the "y" input" -- tests to see if the value x appears in the input field with label y

I fill in the basic work information for "[title]" -- sticks some default values in for all the fields in the post new work form.