Okay, trying again! This time with really fixed sorting-by-size of collections (sorts based on total items including in subcollections, not just in top-level) and hopefully all filters and sorting working. :P
Also fixed sanitizer config so it is now consistent with the allowed_html instructions. The following previously-cut tags should now be allowed through: 

acronym, address, dfn, ins, kbd, samp, var

The help should now also list the "s" tag which was previously allowed but not listed, and should NOT list the datetime attribute, which is not supported by any major browser, so have removed it rather than add to the config.