Adding ids to fixtures to make it possible for sphinx to index them reasonably quickly.
This does make the fixtures harder to read, but we run the tests a whole lot more often than we edit the fixtures. The problem is that if you don't provide an id, Rails creates one based on the fixture name, and they're large and non-consecutive, which sphinx has trouble with.