Issue 2276: Search results in error if results contain nil values
Checking for nil everywhere we iterate through and display search results, or compacting the values for the cases where the search results structure could be turned into an array.

Have considered the retry_stale option on sphinx, but it seemed like an unnecessary performance load since we reindex often enough anyway.
Nil results could still sneak in, and we'd rather didn't give errors to the user if it can be avoided.