Issue 2441: sometimes trying to approve or remove a related work relationship would give error 500.
Circumstances were right for a double render error to happen; for example, if someone who did not have permissions to do that accessed the page (by typing the link directly, for example.)

redirect_back_or_default has an "and return", but it only applied to that function in itself, and adding another "and return" to that line didn't do the trick.

Added a scenario that's not directly related, and a commented scenario for the visible drafts issue.