this doesn't close issue 2495, it's just a first pass
backgrounded reading saves and running them serially in order to
keep a burst of readings from all hitting the database at the same
time.

As with the hit counters, it would probably be better to run both of
them continuously: the 1 hour rake process is just a placeholder
for now until we see how the performance is.