Issue 2310 - reading major_version_read can be nil, which causes error 500 on the user's reading page
This is a bit of a hot fix, but reading the code, major_version_read
should never be nil, so it doesn't seem to make sense to add a special
if case for nil. Only checking if it's different than the work's major
version seems sufficient - the work's major_version will only go up,
not down, so the reading's major version can either be the same, or
lower. Or nil.