Fixes for failing cukes - series word count
Pick one! The cukes were failing, most likely, because of a nil
somewhere, which is why .compact solves it (when the summing is made
through the array), and summing through the database also solves it.

Which is more performant? We can't be entirely certain; it might be
faster with arrays, since there won't ever be enormous number of
records?