more test fixes
Several of these are attempts  to fix errors that show up in Travis that I can't reproduce -- the solutions run clean so not likely to be worse. Also fixed several of the "show me the" generic steps helpful for debugging. Moved collectible.feature which belongs in collections/ rather than other/ too.

Signed-off-by: shalott <shalott@gmail.com>