Hey, did you know automated testing is cool?
Thanks to having a way to test it, I found the proper way to expire controller caches, so guest kudos show up immediately!

Bad news, however, are that there's no clean way to expire the cache for kudos when the pseud used on the kudo has changed, whether by being replaced with a new default pseud because of pseud deletion, or by being removed entirely due to user account deletion.