Fixes AO3-3720, pseuds not updating in autocomplete
With shiny tests, yay. The issue was (1) the removal/addition wasn't being triggered at all because the condition check was looking for :name instead of "name" etc, and (2) the removal wasn't using the correct stale values for pseud/username, and (3) for username changes, the correct current value had to be reloaded from the database before adding.

Signed-off-by: astolat shalott <shalott@gmail.com>