A few more tweaks to comment moderation
approve button in inbox is fixed but as this cannot work without javascript, without javascript the user will instead see "Unreviewed Comments" and that will be a link to the unreviewed comments page where they can approve.
NOTE: the javascript behavior can't be tested automatically.

After posting a moderated comment:

if you are logged in, you will now be redirected to the comment page itself where you can edit/delete it

if you are a guest, you will be scrolled down to the general comments section and will see the "will appear publicly after approved" notice

inbox comments will be marked read when the corresponding comment is approved, regardless of when/where it is approved. This should update automatically on the homepage and the inbox.

The missing blank line in comment notification email should be fixed.

Signed-off-by: astolat shalott <shalott@gmail.com>