AO3-4108 & AO3-4367 Moved author reveal emails into CreationObserver.
This makes it possible to access updated values of in_anon_collection and
in_unrevealed_collection during the before_update callback for Works, so
that it's easier to check whether an item is in two anonymous collections
when doing reveals (fixing bug AO3-4108). As a happy accident, this also
fixes AO3-4367 (the bug where author reveals aren't sent out for a
collection that goes straight from unrevealed to revealed, with no anon
period).

(This also includes some cucumber tests for AO3-4108, and uncomments the
tests for AO3-4367.)