Removes required param and explicitly includes permitted params
This controller accepts the param challenge_assignments in a manner
similar to, but not, nested attributes. This set of strong params does
not require a `challenge_assignments` param to be present, but does
give protection by explicitly identifying parameters allowable to be
consumed by this controller. Eventually, it would be preferable to
investigate switching to a more updated nested attribute refactor to
allow the traditional `require` block to be included in the strong
params definition.