Adds base rails params to challenge_assignments_controller
This controller applies strong params to an implementation that mimics
nested attributes. Because the `permit` method is called on the base
params vs scoped to the model params, I am passing through the
expected root level params as well. Eventually, this is a good
candidate to refactor to standard rails nested params potentially.