AO3-4873 Implements strong parameters for Tag Set Nominations (#2763)
* AO3-4873 Implements strong parameters for Tag Set Nominations

* Adds empty param under tag_set_nomination to avoid being stripped

With the addition of strong params, the tag_set_nomination param is
required. The spec was passing an empty hash which gets stripped and
results in the param not getting passed to strong params and resulting
in a 400 error. Adding an empty nested param (such as pseud_id) keeps
the tag_set_nomination param in the request while avoiding passing any
actual modifying data to the test.

For referehce, see issue: https://github.com/rails/rails/issues/26569

* Remove leftover spec focus

* AO3-4873 Adds updated params based on @tickinginstant's QA