AO3 4861 challenge assignment strong params (#2764)
* AO3-4861 Implements strong params and enables ForbiddenAttributes on ChallengeAssignment

* Removes required param and explicitly includes permitted params

This controller accepts the param challenge_assignments in a manner
similar to, but not, nested attributes. This set of strong params does
not require a `challenge_assignments` param to be present, but does
give protection by explicitly identifying parameters allowable to be
consumed by this controller. Eventually, it would be preferable to
investigate switching to a more updated nested attribute refactor to
allow the traditional `require` block to be included in the strong
params definition.

* Adds base rails params to challenge_assignments_controller

This controller applies strong params to an implementation that mimics
nested attributes. Because the `permit` method is called on the base
params vs scoped to the model params, I am passing through the
expected root level params as well. Eventually, this is a good
candidate to refactor to standard rails nested params potentially.