AO3-4954 Check for empty params when updating tags
It's possible for an edit form to send back just the tag type
and the fix taggings count (e.g. changing the type of an unsorted tag),
leaving the controller to enforce strong parameters on an empty hash:

    "Required parameter missing: tag"

To avoid this, we only apply strong parameters if params is not empty.