AO3-3752 Stop concatenating author attributes
We can't concatenate params[:chapter][:author_attributes][:ids] and params[:chapter][:author_attributes] because strong parameters aren't mutable. This code was no longer working, which was a good thing because it was originally responsible for every existing co-author being added to new chapters of a work, regardless of whether you wanted that.