AO3-3176 Prevent 500 error when editing chapter you didn't coauthor (#2627)
* AO3-3752 Let user posting a chapter set the co-creators

* AO3-3752 Don't add co-creators by default

* AO3-3176 Have to use type rather than @chapter to avoid affecting work form

* AO3-3752 Only show co-authors when there are co-authors

* AO3-3752 Tests for co-authored chapters, co-author test step changes

* AO3-3752 Make the step for posting (draft) chapters work again

* AO3-3752 Fix spacing in chapter edit tests

* AO3-3176 Add test showing you can now edit a chapter you are not a co-author of

* AO3-3176 Add steps to prove we are editing the right chapter

* AO3-3752 Remove test for unrelated issue as this no longer fixes it

* AO3-3752 Use byline format for cocreator checkbox labels

* AO3-3752 Use double quotes in the view

* AO3-3752 Use [author_attributes][ids] rather than [coauthors] because strong parameters

We originally stuffed the coauthors params into the ids params in the controller, but we can't do that anymore with strong parameters.  We can possibly eliminate the stuffing step and use ids to begin with.

* AO3-3752 Stop concatenating author attributes

We can't concatenate params[:chapter][:author_attributes][:ids] and params[:chapter][:author_attributes] because strong parameters aren't mutable. This code was no longer working, which was a good thing because it was originally responsible for every existing co-author being added to new chapters of a work, regardless of whether you wanted that.

* AO3-3752 Fix mistake from merging with master

* AO3-35752 More tests relating to chapter coauthoring

* AO3-3752 Remove coauthor field from hidden fields and don't allow the parameter

* AO3-3752 ACTUALLY remove coauthors for chapters allowed parameters

* AO3-3176 Let any work co-author edit any chapter of the work

AO3-3176 Add editing user as a co-author to the chapter they are editing

AO3-3176 Don't override selected_pseuds in the view

AO3-3176 Consolidate the logic in the pseuds helper

AO3-3176 Better name for helper method

* AO3-3176 Don't add more unnecssary interger conversion for pseud.id, and fix a line break in a code comment