Fixes problem with includes statement
The includes statements in question referenced a has_many association that used
conditions (approved_conditions) that applied to an association of the
association. (`Work.includes(:approved_collections)` where `approved_collections`
are conditions that have collections that meet a certain criteria). By
moving the scopes that determine whether a collection is
approved/rejected/approved by user into the collectible model, the Work model
doesn't have to guess what is meant by the reference to the columns
`collection_items.attributes`.