Fixes bug in collection_item after_commit callback
In Rails 4.2 and before, ActiveRecord suppressed errors raised within callbacks.
Therefore it was suppressing a case in the collection_item after_commit callback
being called when a collection has been destroyed. The callback reference's the
collection's collection_preference object and calls a method on it
(email_notify). The collection_preference record has been destroyed at this
point. In 4.2 and before, AR suppressed the error and moved on. Since it's
happening in an if statement, it worked effectively to not send out the
notification and treat the error as a false case.

This PR just explicitly spells out the false case so that an error isn't raised,
since it will not be suppressed.