Prevents error in work callback
Weird problem.

@work.destroy is triggering the *_save callbacks when running:

`RAILS_ENV=test bundle exec rspec spec/lib/works_owner_spec.rb[1:4:2:1:2]`

(when the work's owner is a collection)

This is happening on master. I'd consider it unexpected behavior, but it doesn't
seem to have any side-effects. The only reason it's causing an error now is
because Rails 5 no longer swallows errors raised in callbacks (in this case,
undefined method published_at for nil, because the work's chapters have already
been deleted by the time the *_save callbacks are re-triggered).