AO3-3223 When saving a coauthor for a work, save for all existing chapters as well (#2922)
* When saving a coauthor for a work, save for all existing chapters as well

The current behavior is that when a coauthor is saved for a work, they are
only saved as a coauthor for the first chapter. This change makes it so that
they are saved as a coauthor for every chapter on the work. However, it does
not change it so that each new chapter contains all the work's coauthors;
they will only have the user who saved that chapter as an author, and all
other coauthors have to be added manually.

* Changes for the hound

* Fixing orphan work test, which relied on adding a coauthor to only add the coauthor to the first chapter
Adding a test to make sure byline appears on all chapters

* Updating cukes to adhere to convention (and updated user creation)