A03-3536 Show validation error message when parent Url unreachable (#2954)
When an archive work is associated with an external work that has
an unreachable URL, the previous behavior was to display a generic
"Parent work couldn't be saved" message.  Now we will instead show
the first validation error message for that external work, which
should provide enough information for the user to deal with the
underlying problem.