AO3-5305 Fix new-search feature tests for Codeship
The new UI is only enabled if ES 0.90 is not running. This is not
a problem for the Travis build where there's only one active ES
version per job, and new/old-search-only tests are disabled
accordingly.

In the Codeship build, both ES versions are active, which means
the new UI is not available. We need to enable it specifically for
the tests tagged new-search.