AO3-3880 Fix table of contents generation for mobi and epub
* Override Calibre's default XPath for detecting chapters
* Add invisible h2s to HTML to set TOC section names

Note: The previous version of downloads did not generate a
proper TOC for mobi files -- it merely placed a list at the
top of the document. Now it has a proper TOC like epubs
always had.

Also note that everything starting with the h1 in
download_preface is now on a separate page in epub and mobi,
but is still considered part of the preface. The separate
happens because Calibre auto inserts them; this can be
overridden but was beyond the scope of this commit.