AO3-5465 Modify the crossover check. (#3405)
Currently, two fandoms are considered "related" for the purposes of the
crossover check if one fandom's first-class meta tags are a subset of
the other fandom's first-class meta tags. This means that if you have a
fandom A with meta tags C and D, and a fandom B with meta tags D and E,
then any work tagged with A and B counts as a crossover (because A has a
meta tag C that isn't shared by B, and B has a meta tag E that isn't
shared by A).

This commit modifies the crossover check so that two fandoms count as
"related" if they share a meta tag, meaning that fandoms A and B would
count as related, and works tagged with both won't automatically count
as crossovers.