AO3-5631 Make /tmp download directory unique per generation
Add the file type and current timestamp to the directory name.
This means each generation has to stay with the same Download
instance for proper cleanup, and the Download class methods
become obsolete.

Having timestamped directory names means we need to freeze time
so the tests can find the generated files.