Make contact resilient to name changes
The `lastName` field is required by Zoho. It's the only required field for a contact, in fact.

As we, of course, don't ask for people's last names, we are content to use the email, because that's the field that *we* require so we know it's going to be present.

Furthermore, we don't want to store the name (which is *not* necessarily the username for a logged in Archive user) with the contact, because people might be giving different names for different tickets.

We will introduce a custom name field for the ticket in the near future.