Use a custom field `Name` for ticket-level names
Users might enter different names, or no names, for tickets they send with the same email address.

The email address is the unique identifier for a contact, so we don't want to rely on Zoho's automated lookup of the `contactName`.

Instead, we want to give the user the chance to tell us what name they want to be used in association with this specific ticket.