AO3-5480 Switch to Zoho Desk REST API for Support and Abuse forms (#3605)
* (fix) Typo in regex

The typo meant we were too lenient with some blank spaces, but removing any incidental `s` after the starting paragraph mark.

* (chore) Implement mini Zoho clients

A client for authorisation and one for the "resources", since this is how Zoho separate their concepts (different servers, endpoints etc).

* (chore) Feedback reporters use the new Zoho API

* Doh

Thanks, Github Desktop, I hate it.

* Update default config.yml

* Fix Hound comments

* Move specs

* Delete templates

* Use ArchiveConfig to allow stubbing

* Add ArchiveConfig.ZOHO_REDIRECT_URI

* Encode query parameters of Zoho GET requests

* Add comments for new Zoho config keys

* Fall back on Language.default

* Fix Hound comments on specs

* Don't need both `query:` and `.to_query`.

* Fix broken specs.

* Cache the access token.

* (chore) Obey the Hound and the sarken

The double quotes were sarken's request, all the rest was Hound's hounding.

* (fix) Indentation

* Make contact resilient to name changes

The `lastName` field is required by Zoho. It's the only required field for a contact, in fact.

As we, of course, don't ask for people's last names, we are content to use the email, because that's the field that *we* require so we know it's going to be present.

Furthermore, we don't want to store the name (which is *not* necessarily the username for a logged in Archive user) with the contact, because people might be giving different names for different tickets.

We will introduce a custom name field for the ticket in the near future.

* Use a custom field `Name` for ticket-level names

Users might enter different names, or no names, for tickets they send with the same email address.

The email address is the unique identifier for a contact, so we don't want to rely on Zoho's automated lookup of the `contactName`.

Instead, we want to give the user the chance to tell us what name they want to be used in association with this specific ticket.

* Zoho new custom fields format

A very recent change! `customFields` is now deprecated, replaced with the much clearer and self-explanatory `cf`.</sarcasm>

Also, the custom fields themselves must be sent as `cf_snake_case_name`, even though in the interface you can only give the display name. They automatically compute the snake case form from the display name, so `Archive Version` becomes `cf_archive_version` and so on.

* AO3-5480 Change email veracity check to format check

We don't want to allow email addresses with asterisks because Zoho doesn't let contacts have emails like that.

* AO3-5480 Update support and abuse form tests to check for proper error message