(fix) Check the current user is a maintainer before showing privileged info
Previously, we were assuming that the `@collection` variable being set means a mod is the user currently looking at the page.

This assumption turned out to be erroneous, because `@collection` is also set when a non-maintainer user has multiple assignments in the same challenge.

The reason this happens for users with multiple assignments is that the challenge sidebar sends the user to a different path when that's the case:
```
    <% elsif @challenge_assignments.size > 1 %>
      <li><%= span_if_current ts("My Assignments"), user_assignments_path(current_user, :collection_id => @collection.name) %>
    <% end %>
```

This goes to the `index` action of the `challenge_assignments_controller`, where this happens:
```
  def index
    if params[:user_id] && (@user = User.find_by(login: params[:user_id]))
      if current_user == @user
        if params[:collection_id] && (@collection = Collection.find_by(name: params[:collection_id]))
```
and `@collection` gets assigned.