Use the factory_bot gems instead of factory_girl
The gem previously known as `factory_girl` (with its associated Rails specific `factory_girl_rails`) is now `factory_bot` (and `factory_bot_rails` respectively), so any version updates will happen to the ’`bot`.

It’s good practice to have the latest stable versions of gems if they’re compatible with our code without too much fuss.

I’ve used the official migration guide:
https://github.com/thoughtbot/factory_bot/blob/4-9-0-stable/UPGRADE_FROM_FACTORY_GIRL.md

Attempting to jump to the latest version (5.0.2 at the time of writing) had too many failures, so I'm now taking it in steps.

The first step is as close to a drop-in replacement as possible, by locking `factory_bot` at 4.8.2.

We still needed to update pickle at the same time, or the tests would fail because of some nonsense with virtual attributes such as `terms_of_service` and `age_over_13` in user factories.