AO3-5639 Don't count hits on unrevealed and hidden works
Previously, visiting a work that was part of an unrevealed collection was increasing the hit count, even though all a regular (non-admin) visitor could access was a message about the work being unrevealed, and not the work itself. This was upsetting or spooking out the creators, because "unrevealed" suggests no one can access the work other than its creator(s).

This commit adds a check that the work is not part of an unrevealed collection.

Previously, visiting a work that was hidden by an admin (which all admins can do) was increasing the hit count. This was upsetting to the creators, because "hidden" suggests no one can access the work other than its creator(s).

This commit adds a check that the work is not hidden by admin.

I've also consolidated the checks that prevent a hit count from being recorded into one method, to make the code more readable.

We have briefly discussed disregarding *all* admins visits from hit counts, but decided this was out of scope for this issue, and we do not feel strongly enough about this to allow the scope creep.